rename sensors lib
