+++ src/components/SensorServiceIcons/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {List} from 'react-native-paper';
-import {bleCharacteristicToIconName} from '../../lib/sensors';
+import {bleCharacteristicToIconName} from '../../lib/sensor';
 
 type Props = {
   data: string[];
