+++ src/lib/sensor.ts
