+++ src/screens/Sensors/index.test.tsx
@@ -7,10 +7,9 @@ import {Button, Text} from 'react-native-paper';
 
 import App from '../../../App';
 import {navigate} from '../../lib/navigation';
-import {ble} from '../../lib/sensors';
 import Sensors from './';
 import {db, SensorModel} from '../../database';
-import {createSensor, getAllSensors} from '../../lib/sensors';
+import {ble, createSensor, getAllSensors} from '../../lib/sensor';
 import {Sensor} from '../../components/Sensor';
 
 let screen!: ReactTestInstance;
